requirements.txt
# VNPT Money GraphRAG Chatbot - Requirements
# Python 3.10+

# ===========================================
# Core Framework
# ===========================================
chainlit>=1.0.0
python-dotenv>=1.0.0

# ===========================================
# LLM & Embeddings
# ===========================================
openai>=1.12.0

# ===========================================
# Database
# ===========================================
neo4j>=5.15.0
redis>=5.0.0

# ===========================================
# Data Processing
# ===========================================
numpy>=1.24.0
pandas>=2.0.0

# ===========================================
# Utilities
# ===========================================
pydantic>=2.5.0
tenacity>=8.2.0

# ===========================================
# API & Dashboard
# ===========================================
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0

# ===========================================
# Evaluation (optional)
# ===========================================
# ragas>=0.1.0
# datasets>=2.14.0

# ===========================================
# Testing (optional)
# ===========================================
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ===========================================
# Development (optional)
# ===========================================
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
